requirements.txt
azure-core==1.31.0
azure-functions==1.21.3
azure-storage-blob==12.23.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
distlib==0.3.6
filelock==3.12.0
google-analytics-data==0.18.12
google-api-core==2.21.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
gunicorn==20.1.0
idna==3.10
isodate==0.7.2
oauthlib==3.2.2
platformdirs==3.5.1
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.23.0
pytest==8.3.3